Allow for cssClass and code on resolveMenuItem
It would be great, if one could attach cssClass (and/or code) to MenuItems that are dynamically generated by a custom Plugin via the `pages.menuitem.resolveItem`-Event.

I understand `code` and `cssClass` should actually be under control of the Plugin user/developer, but otherwise there is no way, to attach e.g. classes to these dynamic Items, since they don't show up in the Menu editor of the PagesPlugin.