Slugify page URLs before turning them into filenames.
Reverts #420 by replacing with #416's original solution. Credit to @FlusherDock1 for the solution and @acasar for the report of the fix being broken. The PageList class has too much code that relies on baseFileName comparisons to have the fileName switch make sense, especially when factoring in the thousands of existing sites storing baseFileName in their configs instead of fileName.