Populate layout on AJAX requests when editing new page
This fixes some "widget not bound to controller" errors appearing under the following scenario:

When a new page form is opened, the page is populated with the default layout as per the "staticPage" component's "Default Layout" checkbox. However, the selected layout value is not populated in post (as the field doesn't change), so on any subsequent AJAX requests, the layout for the new page is blank. Thus, the "getLayoutObject" method in the Page class returns the first layout in the list, which may or may not be your default layout - if it is not, the form widget is loaded with the wrong layout's fields, which causes the widget not bound errors.

Fixes https://github.com/octobercms/october/issues/4628